requirements.txt
Django==1.8.2
Pillow==2.8.1

